Refactored main widgets and improved project structure.
